Test scrip added, was toying with using a DB to store the profiles in, but ran into issues so reverted back to the normal way of placing a file in UDATA. (seems to work fine the old way)
Also updated the scrip to not load the menuLoader if its not meant to & also updated the progress dialogs & startup.xmls for the skins.

Now when startup.xml is loaded & the individual saves setting is enabled it wont load the games section or home menu.
(how it used to be before i started arsing about)