Renamed the images for the scripts GUI (will also update them at some point)
Fixed the generate favourites script, and added an info section and the option to use your games directories.
(also made it so if there is not game tbn it uses the XBMC cached one)

Added a modules check for the 480p Game Loader script as these modules are required for it to function.