New view, CardView.
Updated all the views that have the trainer and played icon overlays:
If the game hasnt been played yet the trainer icon will move over into its place, makes it cleaner looking.


Updated the fanart view, added the played and trainer overlays.

Updated some of the textures ( larger dimensions ) and also optimized most of the images.
( Super PNG )

Updated the Synopsis and the video preview screen:
Fanart and the background overlay are delays, this stops the flickering effect.

Removed some includes from the includes.xml as there only used on the master profiles loginscreen.

Created a new birthday xml that I wanted to use and add settings so people could set up to 5 birthdays for your kids and on that day it would have balloons and a banner.
Well since we cannot SubSting info labels its cannot be done, I did create a workaround but it slowed XBMC down alot lol nearly 2000 lines of code to do what 1 1 line should be able to do and that's only the non American and Canada day month format, its another 1900 lines for that one lol.
( Includes_SubString_Workaround.xml is not used )